+++ src/compute-engine/latex-syntax/dictionary/definitions-complex.ts
@@ -1,0 +1,24 @@
+import { LatexDictionary } from '../public';
+
+export const DEFINITIONS_COMPLEX: LatexDictionary = [
+  {
+    name: 'Real',
+    kind: 'function',
+    latexTrigger: ['\\Re'],
+  },
+  {
+    name: 'Imaginary',
+    kind: 'function',
+    latexTrigger: ['\\Im'],
+  },
+  {
+    name: 'Argument',
+    kind: 'function',
+    latexTrigger: ['\\arg'],
+  },
+  {
+    name: 'Conjugate',
+    latexTrigger: ['^', '\\star'],
+    kind: 'postfix',
+  },
+];