+++ src/compute-engine/latex-syntax/public.ts
@@ -43,6 +43,7 @@ export type LibraryCategory =
   | 'collections'
   | 'control-structures'
   | 'combinatorics'
+  | 'complex'
   | 'core'
   | 'data-structures'
   | 'dimensions'